requirements.txt
streamlit==1.32.0
pandas==2.2.1
openpyxl==3.1.2
python-telegram-bot==20.8 